fix(acp): address review issues from PR #2142
- Extract shared resolveAcpDisplayName helper to eliminate 3x duplicated displayName logic
- Move formatAcpLogEntry from AcpLogsPanel.tsx to dedicated acpLogFormatter.ts utility
- Add runtime validation for JSON.parse result in useAcpInitialMessage
- Guard markWarmupReadyStatus against stale warmup overwriting terminal status
- Clean up module-level recovery UI state on conversation deletion
- Prefix unused kill() reason parameter to satisfy lint
- Replace inline marginBottom style with UnoCSS mb-12px class in banners
- Prefix unused correctedVersion variable in acpConnectors

Review follow-up for #2142
