+++ src/renderer/pages/conversation/platforms/acp/AcpConnectionBanner.tsx
@@ -1,4 +1,4 @@
-import { ACP_BACKENDS_ALL } from '@/common/types/acpTypes';
+import { resolveAcpDisplayName } from '@/renderer/pages/conversation/utils/resolveConversationBackend';
 import { Alert, Button, Space, Typography } from '@arco-design/web-react';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
@@ -15,10 +15,7 @@ type AcpConnectionBannerProps = {
 const AcpConnectionBanner: React.FC<AcpConnectionBannerProps> = ({ agentName, backend, retrying = false, onRetry }) => {
   const { t } = useTranslation();
 
-  const displayName =
-    agentName ||
-    ACP_BACKENDS_ALL[backend as keyof typeof ACP_BACKENDS_ALL]?.name ||
-    backend.charAt(0).toUpperCase() + backend.slice(1);
+  const displayName = resolveAcpDisplayName(backend, agentName);
 
   return (
     <Alert
@@ -36,7 +33,7 @@ const AcpConnectionBanner: React.FC<AcpConnectionBannerProps> = ({ agentName, ba
           </Space>
         </Space>
       }
-      style={{ marginBottom: 12 }}
+      className='mb-12px'
     />
   );
 };
